+++ .circleci/config.yml
@@ -72,6 +72,17 @@ jobs:
           paths:
             - ~/.npm
 
+  comment_on_pr:
+    <<: *container_config
+    steps:
+      - *restore_repo
+      - *sum_lockfiles
+      - *restore_npm_cache
+      - *npm_install
+      - run:
+        name: Comment on PR
+        command: npm run githubComment
+
   check_for_manual_release:
     <<: *container_config
     steps:
@@ -186,6 +197,9 @@ workflows:
     jobs:
       - checkout_code
       - install_dependencies:
+          requires:
+            - checkout_code
+      - comment_on_pr:
           requires:
             - checkout_code
       - check_for_manual_release:
