+++ scripts/githubComment.js
@@ -1,0 +1,9 @@
+/* eslint-env node */
+/* eslint-disable @typescript-eslint/no-var-requires */
+const { getAllPublicPackages } = require("./getPackages");
+
+function githubComment() {
+  console.log(getAllPublicPackages());
+}
+
+githubComment();
